Changes to candidate selection to address selecting double-check work
git-svn-id: https://svn.code.sf.net/p/prpnet/code@72 029e2237-403a-42e8-9450-7768c132b01b

+++ source/LengthCalculator.cpp
@@ -191,7 +191,7 @@ void     LengthCalculator::CalculateDecimalLengths(Socket *theSocket)
          {
             case ST_GENERIC:
                // Default the length.  The client will compute a more accurate length.
-               decimalLength = strlen(candidateName);
+               decimalLength = (double) strlen(candidateName);
                break;
 
             case ST_FACTORIAL:
